Move `WelcomeBox` and `MainBox` to app.py in one_window example
